+++ mocha-environment.js
@@ -1,5 +1,8 @@
+/* eslint prefer-template: 0 */
+/* eslint object-shorthand: 0 */
+
 // import es6
-require('babel-core/register');
+require('babel-register');
 
 // jsdom
 var jsdom = require('jsdom'); // eslint-disable-line
@@ -9,10 +12,7 @@ global.window = global.document.defaultView;
 global.navigator = global.window.navigator;
 
 global.window.URL = {
-  createObjectURL: function (arg) { // eslint-disable-line
+  createObjectURL: function createObjectURL(arg) {
     return 'data://' + arg.name;
   }
 };
-global.window.addEventListener('load', function () {
-  console.log('JSDOM Loaded'); // eslint-disable-line
-});